+++ ndk-sys/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Generate against upstream NDK build `11769913`. (#471)
 - Add `nativewindow` feature to link against `libnativewindow`. (#465)
+- Add `AIBinder` and `Parcel`
 
 # 0.5.0 (2023-10-15)